# openapi/ema.v1.s16a.json
{
...
  "info": {
    "title": "Ema local API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/reporting/runs": {
      "post": {
        "operationId": "handler_reporting_runs_post",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "client_ids": [
                  "client-exemplu"
                ],
                "years": [
                  2023,
                  2024,
                  2025
                ]
              },
              "schema": {
                "properties": {
                  "client_ids": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "years": {
                    "items": {
                      "type": "integer"
                    },
                    "type": "array"
                  }
                },
                "required": [
                  "years",
                  "client_ids"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "client_ids": [
                    "client-exemplu"
                  ],
                  "exceptions": [],
                  "id": "run-exemplu",
                  "output_id": "output-exemplu",
                  "years": [
                    2023,
                    2024,
                    2025
                  ]
                },
                "schema": {
                  "properties": {
                    "client_ids": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "exceptions": {
                      "items": {},
                      "type": "array"
                    },
                    "id": {
                      "type": "string"
                    },
                    "output_id": {
                      "type": "string"
                    },
                    "years": {
                      "items": {
                        "type": "integer"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "id",
                    "years",
                    "client_ids",
                    "exceptions",
                    "output_id"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      }
    },
    "/reporting/runs/{run_id}": {
      "get": {
        "operationId": "handler_reporting_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "client_ids": [
                    "client-exemplu"
                  ],
                  "exceptions": [],
                  "id": "run-exemplu",
                  "output_id": "output-exemplu",
                  "years": [
                    2023,
                    2024,
                    2025
                  ]
                },
                "schema": {
                  "properties": {
                    "client_ids": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "exceptions": {
                      "items": {},
                      "type": "array"
                    },
                    "id": {
                      "type": "string"
                    },
                    "output_id": {
                      "type": "string"
                    },
                    "years": {
                      "items": {
                        "type": "integer"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "id",
                    "years",
                    "client_ids",
                    "exceptions",
                    "output_id"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      }
    },
...
  }
}